fix: many components fixes;
