+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "hotstaq",
-  "version": "0.8.8",
+  "version": "0.8.9",
   "description": "A friendly web framework that fits nicely into devops and CI/CD pipelines.",
   "bin": {
     "hotstaq": "./bin/hotstaq"
