+++ src/HotStaq.ts
@@ -21,6 +21,9 @@ import { HotTestDestination} from "./HotTestDestination";
 
 import { HotSite, HotSiteRoute } from "./HotSite";
 
+import { registerComponent } from "./HotStaqRegisterComponent";
+import { hotStaqWebStart } from "./HotStaqWebStart";
+
 var HotTesterMocha: any = null;
 var HotTesterMochaSelenium: any = null;
 var HotTestSeleniumDriver: any = null;
@@ -124,7 +127,7 @@ export class HotStaq implements IHotStaq
 	/**
 	 * The current version of HotStaq.
 	 */
-	static version: string = "0.8.8";
+	static version: string = "0.8.9";
 	/**
 	 * Indicates if this is a web build.
 	 */
@@ -535,13 +538,13 @@ export class HotStaq implements IHotStaq
 			throw new Error (`Component ${tempComponentObj.tag} already exists!`);
 
 		this.components[tempComponentObj.tag] = { componentType: ComponentType, processor: this, api: tempApi };
-		this.registerComponent (tempComponentObj.tag, elementOptions);
+		registerComponent.call (this, tempComponentObj.tag, elementOptions);
 	}
 
 	/**
 	 * Correct any HTML prior to DOM parsing. This only accounts for <tr> currently.
 	 */
-	protected static fixHTML (str: string): { fixedStr: string, querySelector: string; }
+	static fixHTML (str: string): { fixedStr: string, querySelector: string; }
 	{
 		// Take into account the difference between XML and HTML.
 		const tempStr: string = str.replace(/ \/>/g, '>').replace(
@@ -567,379 +570,6 @@ export class HotStaq implements IHotStaq
 		}
 
 		return ({ fixedStr: str, querySelector: querySelector });
-	}
-
-	/**
-	 * Register a component for use as a HTML tag.
-	 */
-	protected registerComponent (tag: string, elementOptions: ElementDefinitionOptions = undefined): void
-	{
-		if ((tag == null) || (tag === ""))
-			throw new Error (`All components must have a tag!`);
-
-		if (customElements.get (tag) !== undefined)
-		{
-			/// @fixme This element has already been defined. Should this throw an error or warning? I don't think it should...
-
-			return;
-		}
-
-		let processorComponents = this.components;
-
-		customElements.define (tag, class extends HTMLElement
-			{
-				/**
-				 * The connected HotComponent.
-				 */
-				component: HotComponent;
-			
-				constructor ()
-				{
-					super ();
-
-					let componentInfo = processorComponents[tag];
-					this.component = new componentInfo.componentType (componentInfo.processor, componentInfo.api);
-				}
-			
-				/**
-				 * This helps parse <tr> and other tags that do not have a parent.
-				 * 
-				 * Thanks Brandon McConnell!
-				 * 
-				 * From: https://stackoverflow.com/questions/67313479/make-parsefromstring-parse-without-validation
-				 * 
-				 * @todo May remove this as it does not seem to work well in a lot of edge cases.
-				 */
-				protected looseParseFromString (parser: DOMParser, str: string) {
-					str = str.replace(/ \/>/g, '>').replace(/(<(area|base|br|col|command|embed|hr|img|input|keygen|link|meta|param|source|track|wbr).*?>)/g, '$1</$2>');
-					const xdom = parser.parseFromString('<xml>'+str+'</xml>', 'text/xml');
-					const hdom = parser.parseFromString('', 'text/html');
-					for (let elem of Array.from(xdom.documentElement.children)) {
-						/// @ts-ignore
-						hdom.body.appendChild(elem);
-					}
-					for (let elem of Array.from(hdom.querySelectorAll('area,base,br,col,command,embed,hr,img,input,keygen,link,meta,param,source,track,wbr'))) {
-						/// @ts-ignore
-						elem.outerHTML = '<'+elem.outerHTML.slice(1).split('<')[0];
-					}
-					return hdom;
-				}
-
-				get observedAttributes(): string[] /// @fixme Does this REALLY have to be static? Awful if it does...
-				{
-					return (this.component.observedAttributes);
-				}
-			
-				async connectedCallback ()
-				{
-					let compHtmlElement = this;
-			
-					// @ts-ignore
-					compHtmlElement.hotComponent = this.component;
-			
-					this.component.htmlElements = [compHtmlElement];
-					this.component.inner = this.innerHTML;
-			
-					if (this.component.handleAttributes != null)
-						await this.component.handleAttributes (this.attributes);
-					else
-					{
-						for (let iIdx = 0; iIdx < this.attributes.length; iIdx++)
-						{
-							const attr: Attr = this.attributes[iIdx];
-							const attrName: string = attr.name.toLowerCase ();
-							const attrValue: string = attr.value;
-			
-							if (attrName === "id")
-								this.component.name = attrValue;
-			
-							if (attrName === "name")
-								this.component.name = attrValue;
-			
-							if (attrName === "value")
-								this.component.value = attrValue;
-			
-							if (attrName.indexOf ("hot-") > -1)
-							{
-								const attrTempName: string = attrName.substring (4);
-			
-								/// @ts-ignore
-								this.component[attrTempName] = attrValue;
-							}
-						}
-					}
-			
-					if (this.component.onPreOutput != null)
-					{
-						if (await this.component.onPreOutput () === false)
-							return;
-					}
-			
-					let outputs = await this.component.output ();
-			
-					if (this.component.onPostOutput != null)
-						outputs = await this.component.onPostOutput (outputs);
-			
-					let componentOutputs: HotComponentOutput[] = [];
-			
-					if (typeof (outputs) === "string")
-						componentOutputs.push ({ html: outputs });
-					else
-					{
-						if (outputs instanceof Array)
-							componentOutputs = outputs;
-						else
-							componentOutputs = [outputs];
-					}
-			
-					for (let iKdx = 0; iKdx < componentOutputs.length; iKdx++)
-					{
-						let output = componentOutputs[iKdx];
-						let htmlStr: string = output.html;
-						let addFunctionsTo: string = "";
-			
-						if (output.addFunctionsTo != null)
-							addFunctionsTo = output.addFunctionsTo;
-			
-						let str: string = HotFile.parseContent (htmlStr, true, { "outputCommands": false });
-			
-						if (this.component.onParsed != null)
-							str = await this.component.onParsed (str);
-			
-						let htmlHandler: { fixedStr: string, querySelector: string; } = { fixedStr: "", querySelector: "" };
-			
-						if (this.component.onFixHTML != null)
-							htmlHandler = await this.component.onFixHTML (str);
-						else
-							htmlHandler = HotStaq.fixHTML (str);
-			
-						let newDOM: Document = null;
-						let newObj: HTMLElement = null;
-			
-						if (this.component.onParseDOM != null)
-							newDOM = await this.component.onParseDOM (htmlHandler.fixedStr);
-						else
-						{
-							/// @ts-ignore
-							//newDOM = this.looseParseFromString (new DOMParser (), str);
-							newDOM = new DOMParser ().parseFromString (htmlHandler.fixedStr, "text/html");
-						}
-			
-						if (newDOM.body.children.length < 1)
-							throw new Error (`No component output from ${this.component.name}`);
-			
-						if (newDOM.body.children.length > 1)
-						{
-							let throwErr: boolean = true;
-			
-							for (let iIdx = 0; iIdx < newDOM.body.children.length; iIdx++)
-							{
-								let child = newDOM.body.children[iIdx];
-			
-								if (child instanceof HTMLElement)
-								{
-									if (child.tagName.toLowerCase () === "parsererror")
-									{
-										newObj = child;
-										throwErr = false;
-			
-										break;
-									}
-								}
-							}
-			
-							if (throwErr === true)
-								throw new Error (`Only a single html element can come from component ${this.component.name}, multiple elements were detected.`);
-						}
-
-						if (htmlHandler.querySelector === "")
-							newObj = (<HTMLElement>newDOM.body.children[0]);
-						else
-							newObj = newDOM.querySelector (htmlHandler.querySelector);
-
-						let childrenToReadd: Node[] = [];
-			
-						// Save the children from being replaced.
-						for (let iIdx = (this.children.length - 1); iIdx > -1; iIdx--)
-						{
-							let child: Node = this.children[iIdx];
-			
-							childrenToReadd.push (this.removeChild (child));
-						}
-			
-						this.replaceWith (newObj);
-			
-						if (this.component.click != null)
-							newObj.addEventListener ("click", this.component.click.bind (this.component));
-			
-						for (let key in this.component.events)
-						{
-							let event = this.component.events[key];
-			
-							// @ts-ignore
-							newObj.addEventListener (event.type, event.func, event.options);
-						}
-			
-						let objectFunctions: string[] = Object.getOwnPropertyNames (this.component.constructor.prototype);
-			
-						// Associate any functions to the newly created element.
-						for (let iIdx = 0; iIdx < objectFunctions.length; iIdx++)
-						{
-							let objFunc: string = objectFunctions[iIdx];
-			
-							if (objFunc === "constructor")
-								continue;
-			
-							// @ts-ignore
-							let prop = this.component[objFunc];
-			
-							if (typeof (prop) === "function")
-							{
-								let isNewFunction: boolean = true;
-			
-								// Go through each function in the base HotComponent and see 
-								// if there's any matches. If there's a match, that means 
-								// we're trying to add an existing function, and we don't
-								// wanna do that. Skip it.
-								for (let key2 in HotComponent.prototype)
-								{
-									if (objFunc === key2)
-									{
-										isNewFunction = false;
-			
-										break;
-									}
-								}
-
-								if (isNewFunction === true)
-								{
-									// @ts-ignore
-									newObj[objFunc] = HotStaq.keepContext (this.component[objFunc], this.component);
-
-									if (addFunctionsTo !== "")
-									{
-										let query: HTMLElement = document.querySelector (addFunctionsTo);
-			
-										// @ts-ignore
-										query[objFunc] = HotStaq.keepContext (this.component[objFunc], this.component);
-									}
-								}
-							}
-						}
-			
-						if (this.component.onPrePlace != null)
-							newObj = await this.component.onPrePlace (newObj);
-			
-						let compHtmlElement2: HTMLElement = await this.component.onCreated (newObj);
-			
-						if (this.component.onParentPlace != null)
-						{
-							// @ts-ignore
-							compHtmlElement2.onParentPlace = this.component.onParentPlace;
-						}
-			
-						// @ts-ignore
-						compHtmlElement2.hotComponent = this.component;
-						this.component.htmlElements.push (compHtmlElement2);
-			
-						if (output.parentSelector != null)
-						{
-							let parentNode: Node = document.querySelector (output.parentSelector);
-			
-							compHtmlElement2.parentElement.removeChild (compHtmlElement2);
-							parentNode.appendChild (compHtmlElement2);
-			
-							// @ts-ignore
-							if (compHtmlElement2.onParentPlace != null)
-							{
-								// @ts-ignore
-								await compHtmlElement2.hotComponent.onParentPlace (parentNode, compHtmlElement2);
-							}
-						}
-			
-						if (output.placeHereParent != null)
-						{
-							let parentNodeToCheck = compHtmlElement2.parentNode;
-							let parentNodeCheckCounter: number = 0;
-			
-							while (parentNodeCheckCounter < 10) /// @todo Make this controllable with a variable from the component.
-							{
-								if (parentNodeToCheck == null)
-									break;
-			
-								if (parentNodeToCheck instanceof HTMLHtmlElement)
-									break;
-			
-								// If the hot-place-here exists, place the children there. If not, place it under the 
-								// new element.
-								let placeHereArray = parentNodeToCheck.querySelectorAll (`hot-place-here[name="${output.placeHereParent}"]`);
-			
-								if (placeHereArray.length > 0)
-								{
-									let placeHere = placeHereArray[0];
-			
-									compHtmlElement2.parentNode.removeChild (compHtmlElement2);
-									placeHere.appendChild (compHtmlElement2);
-			
-									// @ts-ignore
-									if (compHtmlElement2.onParentPlace != null)
-									{
-										// @ts-ignore
-										await compHtmlElement2.hotComponent.onParentPlace (placeHere, compHtmlElement2);
-									}
-			
-									break;
-								}
-			
-								if (placeHereArray.length < 1)
-								{
-									let placeHereAttrArray = parentNodeToCheck.querySelectorAll (`[hot-place-here="${output.placeHereParent}"]`);
-			
-									if (placeHereAttrArray.length > 0)
-									{
-										let placeHere = placeHereAttrArray[0];
-										compHtmlElement2.parentNode.removeChild (compHtmlElement2);
-										placeHere.appendChild (compHtmlElement2);
-			
-										// @ts-ignore
-										if (compHtmlElement2.onParentPlace != null)
-										{
-											// @ts-ignore
-											await compHtmlElement2.hotComponent.onParentPlace (placeHere, compHtmlElement2);
-										}
-			
-										break;
-									}
-								}
-			
-								parentNodeToCheck = parentNodeToCheck.parentNode;
-								parentNodeCheckCounter++;
-							}
-						}
-			
-						// Append the children to the newly created HTML element.
-						for (let iIdx = 0; iIdx < childrenToReadd.length; iIdx++)
-						{
-							const child: Node = childrenToReadd[iIdx];
-			
-							compHtmlElement2.appendChild (child);
-			
-							// @ts-ignore
-							if (child.onParentPlace != null)
-							{
-								// @ts-ignore
-								await child.hotComponent.onParentPlace (compHtmlElement2, child);
-							}
-						}
-			
-						if (this.component.onPostPlace != null)
-						{
-							/// @ts-ignore
-							compHtmlElement2 = await this.component.onPostPlace (compHtmlElement2.parentNode, compHtmlElement2);
-						}
-					}
-				}
-			}, elementOptions);
 	}
 
 	/**
@@ -1231,7 +861,7 @@ export class HotStaq implements IHotStaq
 
 			this.logger.verbose (`Accessing HotSite ${path}`);
 
-			let HotIO = eval ("require")("./HotIO").HotIO;
+			let HotIO = eval ("require")("./HotIO").HotIO; // Hack to get around Webpack.
 			jsonStr = await HotIO.readTextFile (path);
 			this.logger.verbose (`Accessed site ${path}`);
 		}
@@ -1254,7 +884,7 @@ export class HotStaq implements IHotStaq
 
 		const hotsiteStr: string = JSON.stringify (this.hotSite, null, 2);
 
-		let HotIO = eval ("require")("./HotIO").HotIO;
+		let HotIO = eval ("require")("./HotIO").HotIO; // Hack to get around Webpack.
 		await HotIO.writeTextFile (path, hotsiteStr);
 		this.logger.verbose (`Saved site ${path}`);
 	}
@@ -2241,344 +1871,4 @@ hotstaq_isDocumentReady ();
 }
 
 if (typeof (document) !== "undefined")
-{
-	let loadHotStaqSite = function ()
-	{
-		let hotstaqElms = document.getElementsByTagName ("hotstaq");
-
-		// Set this to true, just in case...
-		HotStaq.isWeb = true;
-
-			// @ts-ignore
-		if (typeof (HotStaqWeb) !== "undefined")
-		{
-			// @ts-ignore
-			window.HotStaq = HotStaqWeb.HotStaq;
-			// @ts-ignore
-			window.HotClient = HotStaqWeb.HotClient;
-			// @ts-ignore
-			window.HotAPI = HotStaqWeb.HotAPI;
-			// @ts-ignore
-			window.Hot = HotStaqWeb.Hot;
-			// @ts-ignore
-			window.HotComponent = HotStaqWeb.HotComponent;
-		}
-
-		if (hotstaqElms.length > 0)
-		{
-			// @ts-ignore
-			let hotstaqElm: HTMLElement = hotstaqElms[0];
-
-			setTimeout (async function ()
-				{
-					let getAttr = (elm: HTMLElement, attrNames: string[]) =>
-						{
-							for (let iIdx = 0; iIdx < attrNames.length; iIdx++)
-							{
-								let attrName: string = attrNames[iIdx];
-
-								if (elm.getAttribute (attrName) != null)
-									return (elm.getAttribute (attrName));
-
-								if (elm.getAttribute (`data-${attrName}`) != null)
-									return (elm.getAttribute (`data-${attrName}`));
-							}
-
-							return (undefined);
-						};
-
-					let loadPage: string = getAttr (hotstaqElm, ["load-page", "loadPage", "src"]) || "";
-					let loggingLevel: string = getAttr (hotstaqElm, ["logging-level", "loggingLevel"]) || null;
-					let startDelay: string = getAttr (hotstaqElm, ["start-delay", "startDelay"]) || null;
-					let router: string = getAttr (hotstaqElm, ["router"]) || "";
-					let name: string = getAttr (hotstaqElm, ["name"]) || "default";
-					let args: string = getAttr (hotstaqElm, ["args"]) || null;
-					let apiLibrary: string = getAttr (hotstaqElm, ["api-library", "apiLibrary"]) || null;
-					let apiName: string = getAttr (hotstaqElm, ["api-name", "apiName"]) || null;
-					let apiUrl: string = getAttr (hotstaqElm, ["api-url", "apiUrl"]) || null;
-					let testerName: string = getAttr (hotstaqElm, ["tester-name", "testerName"]) || "HotTesterMochaSelenium";
-					let testerMap: string = getAttr (hotstaqElm, ["tester-map", "testerMap"]) || null;
-					let testerApiBaseUrl: string = getAttr (hotstaqElm, ["tester-api-base-url", "testerApiBaseUrl"]) || null;
-					let testerLaunchpadUrl: string = getAttr (hotstaqElm, ["tester-launchpad-url", "testerLaunchpadUrl"]) || null;
-					let dontReuseProcessor: boolean = false;
-					let passRawUrl: boolean = false;
-					let htmlSource: string = hotstaqElm.innerHTML || "";
-					let routerManager: { [path: string]: { redirect: string; baseRedirect: string; base: string; src: string; } } = {};
-					let routerWildcards: string[] = [];
-					let search: URLSearchParams = new URLSearchParams (window.location.search);
-
-					if (startDelay != null)
-						await HotStaq.wait (parseInt (startDelay));
-
-					if (getAttr (hotstaqElm, ["src"]) != null)
-						loadPage = getAttr (hotstaqElm, ["src"]);
-
-					if (getAttr (hotstaqElm, ["passRawUrl"]) != null)
-						passRawUrl = true;
-
-					if (getAttr (hotstaqElm, ["dont-reuse-processor", "dontReuseProcessor"]) != null)
-						dontReuseProcessor = true;
-
-					let hstqbaseredirect: string = search.get ("hstqbaseredirect");
-
-					if (hstqbaseredirect != null)
-					{
-						hstqbaseredirect = decodeURI (hstqbaseredirect);
-						window.history.replaceState ("", "", hstqbaseredirect);
-						loadPage = hstqbaseredirect;
-					}
-
-					let hotstaqErrors = document.getElementsByTagName ("hotstaq-error");
-
-					for (let iIdx = 0; iIdx < hotstaqErrors.length; iIdx++)
-					{
-						// @ts-ignore
-						let hotstaqErrorElm: HTMLElement = hotstaqErrors[iIdx];
-						let errorStatus: string = getAttr (hotstaqErrorElm, ["status"]);
-						let unsupportedBrowser: string = getAttr (hotstaqErrorElm, ["unsupported-browser-redirect"]);
-
-						if (unsupportedBrowser != null)
-							HotStaq.errors["unsupportedBrowser"] = { redirectToUrl: unsupportedBrowser };
-						else
-							HotStaq.errors[`${errorStatus}`] = { redirectToUrl: unsupportedBrowser };
-					}
-
-					// Check if async/await is available.
-					try
-					{
-						eval ("async () => {}");
-					}
-					catch (ex)
-					{
-						HotStaq.executeError ("unsupportedBrowser");
-					}
-
-					if (router !== "")
-					{
-						let hotstaqRouterElms = document.getElementsByTagName ("hotstaq-router");
-
-						for (let iIdx = 0; iIdx < hotstaqRouterElms.length; iIdx++)
-						{
-							// @ts-ignore
-							let hotstaqRouterElm: HTMLElement = hotstaqRouterElms[iIdx];
-							let routerName: string = getAttr (hotstaqRouterElm, ["name"]);
-							let serveLocally: string = getAttr (hotstaqRouterElm, ["serve-local", "serveLocally"]);
-
-							// @ts-ignore
-							if (routerName === router)
-							{
-								// Load all routes from the router.
-								for (let iJdx = 0; iJdx < hotstaqRouterElm.childNodes.length; iJdx++)
-								{
-									// @ts-ignore
-									let routerElm: HTMLElement = hotstaqRouterElm.childNodes[iJdx];
-
-									if (routerElm instanceof HTMLElement)
-									{
-										if (routerElm.tagName.toUpperCase () === "ROUTE")
-										{
-											let routerPath: string = getAttr (routerElm, ["path"]);
-											let redirect: string = getAttr (routerElm, ["redirect"]);
-											let baseRedirect: string = getAttr (routerElm, ["base-redirect", "baseRedirect"]);
-											let base: string = getAttr (routerElm, ["base"]);
-											let routerSrc: string = getAttr (routerElm, ["src"]);
-
-											if (routerPath.indexOf ("*") > -1)
-												routerWildcards.push (routerPath);
-
-											routerManager[routerPath] = {
-													redirect: redirect || undefined, 
-													baseRedirect: baseRedirect || undefined, 
-													base: base || undefined, 
-													src: routerSrc || undefined
-												};
-										}
-									}
-								}
-
-								let checkPath: string = window.location.pathname;
-								let gotoPath: string = window.location.pathname;
-
-								if (serveLocally != null)
-								{
-									const lowerServeLocally: string = serveLocally.toLowerCase ();
-
-									if ((lowerServeLocally === "true") ||
-										(lowerServeLocally === "yes") ||
-										(lowerServeLocally === "1"))
-									{
-										const lastSlashPos: number = checkPath.lastIndexOf ("/");
-
-										if (lastSlashPos > -1)
-										{
-											checkPath = checkPath.substring (lastSlashPos);
-											gotoPath = gotoPath.substring (lastSlashPos);
-										}
-									}
-								}
-
-								if (routerWildcards.length > 0)
-								{
-									// Serve locally doesn't really work with wildcards
-									/// @fixme This isn't actually working like a wildcard should. This needs to be improved.
-									for (let iJdx = 0; iJdx < routerWildcards.length; iJdx++)
-									{
-										let routeWildcard: string = routerWildcards[iJdx];
-										let tempRouteWildcard: string = routeWildcard.replace ("*", "");
-
-										if (checkPath.indexOf (tempRouteWildcard) > -1)
-										{
-											// This simply returns the key in the routerManager to access.
-											checkPath = routeWildcard;
-
-											break;
-										}
-									}
-								}
-
-								// Find the correct route and load it.
-								if (routerManager[checkPath] != null)
-								{
-									if (routerManager[checkPath].redirect != null)
-									{
-										window.location.href = routerManager[checkPath].redirect;
-
-										return;
-									}
-
-									if (routerManager[checkPath].baseRedirect != null)
-									{
-										const searchParams = window.location.search;
-										let modifiedSearchParams = "";
-
-										if ((searchParams !== "") && (searchParams !== "?"))
-											modifiedSearchParams = `&${searchParams.substring (1)}`;
-
-										window.location.href = `${routerManager[checkPath].baseRedirect}?hstqbaseredirect=${encodeURI (gotoPath + searchParams)}${modifiedSearchParams}`;
-
-										return;
-									}
-
-									if (routerManager[checkPath].src != null)
-										loadPage = routerManager[checkPath].src;
-								}
-
-								break;
-							}
-						}
-					}
-
-					if (args != null)
-						args = JSON.parse (args);
-					else
-						args = Hot.Arguments;
-
-					let hasHtmlSource: boolean = false;
-
-					if (htmlSource !== "")
-					{
-						const htmlSourceCheck: string = htmlSource.replace (/\s/g,'');
-
-						if (htmlSourceCheck !== "")
-							hasHtmlSource = true;
-					}
-
-					let tempMode = 0;
-
-					// @ts-ignore
-					if (window["Hot"] != null)
-						tempMode = Hot.Mode;
-			
-					let processor: HotStaq = null;
-
-					if (dontReuseProcessor === false)
-					{
-						if (typeof (Hot) !== "undefined")
-						{
-							if (Hot.CurrentPage != null)
-							{
-								if (Hot.CurrentPage.processor != null)
-									processor = Hot.CurrentPage.processor;
-							}
-						}
-					}
-
-					if (processor == null)
-						processor = new HotStaq ();
-
-					if (loggingLevel != null)
-						processor.logger.logLevel = HotLog.parse (loggingLevel);
-
-					processor.mode = tempMode;
-
-					let options: HotStartOptions = {
-							name: name,
-							processor: processor,
-							args: args
-						};
-
-					if (loadPage !== "")
-					{
-						if (passRawUrl === false)
-						{
-							if (loadPage.indexOf ("hstqserve") < 0)
-								loadPage += "?hstqserve=nahfam";
-						}
-
-						options.url = loadPage;
-					}
-
-					if (testerMap != null)
-					{
-						options.testerMap = testerMap;
-						options.testerName = testerName;
-					}
-
-					if (testerName != null)
-						options.testerName = testerName;
-
-					if (testerApiBaseUrl != null)
-						options.testerAPIBaseUrl = testerApiBaseUrl;
-
-					if (testerLaunchpadUrl != null)
-						options.testerLaunchpadUrl = testerLaunchpadUrl;
-
-					if (apiName != null)
-					{
-						let client = new HotClient (processor);
-
-						if (apiUrl === "")
-							throw new Error (`api-url was not set!`);
-
-						let parentLib: any = window;
-
-						if (apiLibrary != null)
-						{
-							// @ts-ignore
-							parentLib = window[apiLibrary];
-						}
-
-						let newAPI = new parentLib[apiName] (apiUrl, client);
-						newAPI.connection.api = newAPI;
-						processor.api = newAPI;
-					}
-
-					if (hasHtmlSource === false)
-					{
-						if (loadPage === "")
-							throw new Error (`The hotstaq tag must have a src, HTML contents inside it, or a router set.`);
-
-						HotStaq.displayUrl (options);
-					}
-					else
-					{
-						HotStaq.displayContent (options);
-					}
-				}, 50);
-		}
-	};
-
-	/// @ts-ignore
-	window.ethereum22 = window.ethereum;
-	window.addEventListener ("load", loadHotStaqSite);
-}
+	window.addEventListener ("load", hotStaqWebStart);
