+++ src/api-web.ts
@@ -1,7 +1,7 @@
 import { HotStaq, HotStartOptions, IHotStaq } from "./HotStaq";
 import { HotSite, HotSiteRoute, HotSiteMapPath } from "./HotSite";
 import { Hot, HotModule, HotAsset, DeveloperMode } from "./Hot";
-import { HotComponent, IHotComponent } from "./HotComponent";
+import { HotComponent, HotComponentOutput, IHotComponent } from "./HotComponent";
 import { HotFile } from "./HotFile";
 import { HotLog, HotLogLevel } from "./HotLog";
 import { HotPage } from "./HotPage";
