+++ tests/components/Components.ts
@@ -1,7 +1,14 @@
 import { HelloWorld } from "./HelloWorldComponent";
 import { HelloWorldAPI } from "../server/HelloWorldAPI";
 
+import { TableComponent } from "./TableComponent";
+import { TableHeader } from "./TableHeader";
+import { MainComponent } from "./MainComponent";
+
 export {
 		HelloWorld,
-		HelloWorldAPI
+		HelloWorldAPI,
+		TableComponent,
+		TableHeader,
+		MainComponent
 	};
